# make sure that StartsWith checks only the beginning of the string