@tarekgh 2019-09-05
@adamsitnik I am not sure if last commit is correct. limiting the length can cause some other problems. for example, some Unicode characters don't have sort weights (treated as not exist) or some can ignored with other flags (like ignore symbols). if you limit the search, this will report wrong results.